fixing checkstyle

+++ cv/src/draw_lines.rs
@@ -20,9 +20,9 @@ pub fn draw_lines(mut img: &mut Mat, lines: &Vec<Line>) {
   for l in lines {
     // Truncation here is fine (and needed) as we are just drawing pixels on the screen.
     #[allow(clippy::cast_possible_truncation)]
-    let start_point = Point::new(l.pos1.x as i32, l.pos1.y as i32);
+    let start_point = Point::new(l.pos1.x, l.pos1.y);
     #[allow(clippy::cast_possible_truncation)]
-    let end_point = Point::new(l.pos2.x as i32, l.pos2.y as i32);
+    let end_point = Point::new(l.pos2.x, l.pos2.y);
 
     // OpenCV uses BGR (not RBG).
     let colour = match l.colour {
